entity classes have been moved from to the root `model/` package to keep the structure simple
