+++ netmap-service/src/main/java/com/netmap/netmapservice/model/Company.java
@@ -1,4 +1,4 @@
-package com.netmap.netmapservice.model.skill;
+package com.netmap.netmapservice.model;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -8,12 +8,12 @@ import org.hibernate.type.SqlTypes;
 import java.util.UUID;
 
 @Entity
-@Table(name = "skill")
+@Table(name = "company")
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class Skill {
+public class Company {
 
     @Id
     @GeneratedValue
@@ -22,4 +22,7 @@ public class Skill {
 
     @Column(name = "name", nullable = false, unique = true)
     private String name;
+
+    @Column(name = "description", columnDefinition = "TEXT")
+    private String description;
 }
