+++ netmap-service/src/main/java/com/netmap/netmapservice/model/Skill.java
@@ -1,4 +1,4 @@
-package com.netmap.netmapservice.model.job;
+package com.netmap.netmapservice.model;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -8,12 +8,12 @@ import org.hibernate.type.SqlTypes;
 import java.util.UUID;
 
 @Entity
-@Table(name = "company")
+@Table(name = "skill")
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class Company {
+public class Skill {
 
     @Id
     @GeneratedValue
@@ -22,7 +22,4 @@ public class Company {
 
     @Column(name = "name", nullable = false, unique = true)
     private String name;
-
-    @Column(name = "description", columnDefinition = "TEXT")
-    private String description;
 }
